Revert "Added new dependency for frontend charts"
