+++ zappnode-front/package.json
@@ -6,7 +6,6 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
-    "chart.js": "^2.9.3",
     "classnames": "^2.2.6",
     "primeflex": "^2.0.0-rc.1",
     "primeicons": "^4.0.0",
