+++ zappnode-front/public/data/track-data.json
@@ -1,74 +1,17 @@
 {
   "data": [
     {
-        "uptime": {
-          "data": [
-            65,
-            59,
-            80,
-            81,
-            56,
-            55,
-            40
-          ],
-          "timestamps": [
-            "Monday",
-            "Tuesday",
-            "Wednesday",
-            "Thursday",
-            "Friday",
-            "Saturday",
-            "Sunday"
-          ]
-        },
+        "percentage": 100,
         "name": "CNN",
         "type": "http"
     },
     {
-        "uptime": {
-          "data": [
-            28,
-            48,
-            40,
-            19,
-            86,
-            27,
-            90
-          ],
-          "timestamps": [
-            "Monday",
-            "Tuesday",
-            "Wednesday",
-            "Thursday",
-            "Friday",
-            "Saturday",
-            "Sunday"
-          ]
-        },
+        "percentage": 50,
         "name": "VG247",
         "type": "http"
     },
     {
-        "uptime": {
-          "data": [
-            28,
-            48,
-            30,
-            96,
-            86,
-            75,
-            90
-          ],
-          "timestamps": [
-            "Monday",
-            "Tuesday",
-            "Wednesday",
-            "Thursday",
-            "Friday",
-            "Saturday",
-            "Sunday"
-          ]
-        },
+        "percentage": 100,
         "name": "goal.com",
         "type": "http"
     }
