+++ zappnode-front/src/App.js
@@ -5,17 +5,13 @@ import './tailwind.generated.css';
 import 'primereact/resources/themes/nova-light/theme.css';
 import 'primereact/resources/primereact.min.css';
 import 'primeicons/primeicons.css';
-import 'primeflex/primeflex.css';
 
 import Home from './Home';
 
 function App() {
   return (
     <>
-      <Landing />
-      {/* <Router>
-        <Home />
-      </Router> */}
+      <Landing/>
     </>
   );
 }
