+++ zappnode-front/src/components/Footer.js
@@ -1,15 +1,7 @@
 import React from 'react';
 
 function Footer() {
-  return (
-    <>
-      <footer className="bg-gray-200 text-center bottom-0 text-xs w-full border-t border-gray-400 pin-b px-4 py-1 mt-5">
-        <a href="#">Privacy Policy</a>
-        <a href="#"> | Terms of Service</a>
-        <a href="#"> | Contact Us</a>
-      </footer>
-    </>
-  );
+  return <React.Fragment>Footer</React.Fragment>;
 }
 
 export default Footer;
